Twitch clip same title fix + yt filename fix
Now if a title has the same title as one the clips already downloaded. The script will just put some numbers from the mediaurl into the filename. This prevents the clip already in the folder being deleted.

Also the way youtube thumbnail filenames are chosen has been changed so it picks the whole short code in the link.